requirements.txt
certifi==2022.9.24
charset-normalizer==2.1.1
contourpy==1.0.5
cycler==0.11.0
docopt==0.6.2
fonttools==4.38.0
idna==3.4
kiwisolver==1.4.4
matplotlib==3.6.1
numpy==1.23.4
packaging==21.3
pandas==1.5.1
Pillow==9.3.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.5
requests==2.28.1
six==1.16.0
urllib3==1.26.12
yarg==0.1.9
